// src/app/features/userMaster/components/types.ts
export interface User {
    _id: string;
    userid: string;
    userType: string;
    userName: string;
    email: string;
    mobileNo: string;
    officeName: string;
    officeAddress: string;
    city: string;
    country: string;
  }
  export interface NewUser {
    userid: string;
    userType: string;
    userName: string;
    email: string;
    mobileNo: string;
    officeName: string;
    officeAddress: string;
    city: string;
    country: string;
  }
  
  